Fix tricky trait issue in fb 968
The issue as described in [factory_bot #968][url] is that depending on the order
a parent or child factory is initialized, it will copy over the other's traits
in a way that leads to inconsistent results.

[url]: https://github.com/thoughtbot/factory_bot/issues/968

The fix for us (and hopefully for factory_bot, if I can present this properly)
is to make the "search" for a trait happen every time instead of at compile, and
to stop copying any traits at all into each other. Instead, keep track of the
"parent" (which is effectively a lexical context) inside of each trait, and when
calling "getAttributes" on a trait, have it check up the parent hierarchy for
the existence of a trait with that name.

To handle this, we have to copy over the "parent" of each trait when iterating
over them. This is slightly messy and not my favorite means of handling things,
but because it happens unconditionally, it's okay.

NOTE: You have to delete or move all of the unit tests to run anything, because
typescript is dumb.